gateway error metrics
